Create the backend managers
The managers are powered by a lib which leverages the JobScheduler API or the
GMS included scheduler to perform device radio toggling.

The job scheduler handles multiple job queues, and is reliable to only run an event once.
It makes running jobs in parallel easy, and does not rely on manual threading.

While the inclusion of GCM requires the Internet permission, Power Manager itself does not
access the network in any direct way. The only network access performed would be from the JobManager,
which has the ability to check for network connections