Temporarily disable the back button while onboarding is shown to prevent a crash
When onboarding is shown, back button is still enabled. Pressing it will leave the fragment but the onboarding will stay on screen.
If we are in the middle of a sequence, the onboarding will attempt to show the next sequence item and crash because the view is no
longer alive