Remove sensor restriction code from Doze
The code which restricted sensos was experimental and buggy on many devices. It also, in rare cases, caused
more battery drain than it ended up saving. Because of all of these issues, remove the code which references the
sensor restriction.

An added bonus is we can remove the preference object from the modifiers