Check for root permission at startup
superuser implementations like SU will show a dialog when su is called.
This can lead to the device waking up to display the dialog when it is trying to sleep
as a result of the wakeup, the device may never enter a idle mode

Hopefully, by showing the dialog upon app opening, the user will address Doze and either grant or deny it.
By doing so, we will be able to hopefully dismiss the dialog, allowing the device to properly idle