Special logic for Doze and Airplane mode
Doze and Airplane mode are different from the usual job/manager interface

Normally, the manager turns off the device function if it is on when the screen is off,
and turns it on when the screen is on.

Airplane and Doze are different, they are turned ON when the screen is off, and OFF when
the screen is on. Because of this, they need custom jobs to implement this logic

In the custom job logic, called a ToggleJob instead of a ManageJob, the enable
will actually call the ManageJob.interalDisable(), and vice versa for disable.

Similarly, the managerInteractor will report the opposite of the enabled
state when queiried for a toggle job interface