Logger in base job, fix a bug in manager flow
Because the manager ran under different subscriptions, it could lead to enables and disables
being queued out of order. As a result, it could lead to airplane mode and data for example being
toggled in the wrong order, or while the screen was on.

Subscribe to the same subscription and this should be fixed. We can also add the job in serial order
but on a background thread.