Offload the actual setting commands onto a thread pool
Use a countdown latch to actuall wait for the full job to complete

Ths is done because in the case of root toggles (Data, Airplane, Doze) it may take upwards
of 10 seconds for the full command to complete. While this will not hang the UI, it will
stop any of the other commands from running.

For example, if Wifi waits for Data which waits for Airplane to run, the possible delay between
Wifi actually running may be upwards of 15 seconds.

By threading the commands, all modifiers run in parallel, and the job simply waits for the longest
one to complete using a Countdown latch.

We also don't check for actual SU permission because the command will silent fail anyway, and
SU permission will be granted if the commands actually run